DRY vite configs using a shared config
