+++ packages/shared-components/tsconfig.node.json
@@ -1,6 +1,5 @@
 {
     "compilerOptions": {
-        "composite": true,
         "target": "ESNext",
         "module": "esnext",
         "moduleResolution": "bundler",
