+++ packages/shared-components/vitest.config.ts
@@ -5,57 +5,17 @@ SPDX-License-Identifier: AGPL-3.0-only OR LicenseRef-Element-Commercial
 Please see LICENSE files in the repository root for full details.
 */
 
-import { defineConfig, ViteUserConfig } from "vitest/config";
+import { defineConfig, mergeConfig } from "vitest/config";
 import path from "node:path";
 import { fileURLToPath } from "node:url";
 import { storybookTest } from "@storybook/addon-vitest/vitest-plugin";
 import { storybookVis } from "storybook-addon-vis/vitest-plugin";
 import { playwright, PlaywrightProviderOptions } from "@vitest/browser-playwright";
 import { nodePolyfills } from "vite-plugin-node-polyfills";
-import { Reporter } from "vitest/reporters";
-import { env } from "process";
+
+import baseConfig from "../../vite.shared.js";
 
 const dirname = typeof __dirname !== "undefined" ? __dirname : path.dirname(fileURLToPath(import.meta.url));
-
-const reporters: NonNullable<ViteUserConfig["test"]>["reporters"] = [["default"]];
-const slowTestReporter: Reporter = {
-    onTestRunEnd(testModules, unhandledErrors, reason) {
-        const tests = testModules
-            .flatMap((m) => Array.from(m.children.allTests()))
-            .filter((test) => test.diagnostic()?.slow);
-        tests.sort((x, y) => x.diagnostic()?.duration! - y.diagnostic()?.duration!);
-        tests.reverse();
-        if (tests.length > 0) {
-            console.warn("Slowest 10 tests:");
-        }
-        for (const t of tests.slice(0, 10)) {
-            console.warn(`${t.module.moduleId} > ${t.fullName}: ${t.diagnostic()?.duration.toFixed(0)}ms`);
-        }
-    },
-};
-
-// if we're running under GHA, enable the GHA & Sonar reporters
-if (env["GITHUB_ACTIONS"] !== undefined) {
-    reporters.push([
-        "github-actions",
-        {
-            silent: false,
-        },
-    ]);
-
-    reporters.push([
-        "vitest-sonar-reporter",
-        {
-            outputFile: "coverage/sonar-report.xml",
-            onWritePath: (path) => `packages/shared-components/${path}`,
-        },
-    ]);
-
-    // if we're running against the develop branch, also enable the slow test reporter
-    if (env["GITHUB_REF"] == "refs/heads/develop") {
-        reporters.push(slowTestReporter);
-    }
-}
 
 const commonContextOptions: PlaywrightProviderOptions["contextOptions"] = {
     reducedMotion: "reduce",
@@ -70,95 +30,93 @@ const commonLaunchOptions = {
     args: ["--font-render-hinting=none", "--disable-font-subpixel-positioning", "--disable-lcd-text"],
 };
 
-export default defineConfig({
-    test: {
-        coverage: {
-            provider: "v8",
-            include: ["src/**/*.{ts,tsx}"],
-            exclude: ["src/**/*.stories.tsx"],
-            reporter: [["lcov", { projectRoot: "../../" }]],
-        },
-        reporters,
-        globals: false,
-        pool: "threads",
-        projects: [
-            {
-                extends: true,
-                plugins: [
-                    // The plugin will run tests for the stories defined in your Storybook config
-                    // See options at: https://storybook.js.org/docs/next/writing-tests/integrations/vitest-addon#storybooktest
-                    storybookTest({
-                        configDir: path.join(dirname, ".storybook"),
-                        storybookScript: "storybook --ci",
-                        tags: {
-                            exclude: ["skip-test"],
+export default mergeConfig(
+    baseConfig,
+    defineConfig({
+        test: {
+            coverage: {
+                exclude: ["src/**/*.stories.tsx"],
+            },
+            projects: [
+                {
+                    extends: true,
+                    plugins: [
+                        // The plugin will run tests for the stories defined in your Storybook config
+                        // See options at: https://storybook.js.org/docs/next/writing-tests/integrations/vitest-addon#storybooktest
+                        storybookTest({
+                            configDir: path.join(dirname, ".storybook"),
+                            storybookScript: "storybook --ci",
+                            tags: {
+                                exclude: ["skip-test"],
+                            },
+                        }),
+                        storybookVis({
+                            // 3px of difference allowed before marking as failed
+                            failureThreshold: 3,
+                            // When running in CI=1 mode, set the platform to `linux` as that is the platform where the browser-in-docker is running
+                            snapshotRootDir: ({ ci, platform }) => `__vis__/${ci ? "linux" : platform}`,
+                        }),
+                    ],
+                    test: {
+                        name: "storybook",
+                        browser: {
+                            enabled: true,
+                            headless: true,
+                            provider: playwright({
+                                contextOptions: commonContextOptions,
+                                launchOptions: commonLaunchOptions,
+                                connectOptions: process.env.PW_TEST_CONNECT_WS_ENDPOINT
+                                    ? {
+                                          wsEndpoint: process.env.PW_TEST_CONNECT_WS_ENDPOINT,
+                                          exposeNetwork: "<loopback>",
+                                      }
+                                    : undefined,
+                            }),
+                            instances: [{ browser: "chromium" }],
                         },
-                    }),
-                    storybookVis({
-                        // 3px of difference allowed before marking as failed
-                        failureThreshold: 3,
-                        // When running in CI=1 mode, set the platform to `linux` as that is the platform where the browser-in-docker is running
-                        snapshotRootDir: ({ ci, platform }) => `__vis__/${ci ? "linux" : platform}`,
-                    }),
-                ],
-                test: {
-                    name: "storybook",
-                    browser: {
-                        enabled: true,
-                        headless: true,
-                        provider: playwright({
-                            contextOptions: commonContextOptions,
-                            launchOptions: commonLaunchOptions,
-                            connectOptions: process.env.PW_TEST_CONNECT_WS_ENDPOINT
-                                ? {
-                                      wsEndpoint: process.env.PW_TEST_CONNECT_WS_ENDPOINT,
-                                      exposeNetwork: "<loopback>",
-                                  }
-                                : undefined,
-                        }),
-                        instances: [{ browser: "chromium" }],
-                    },
-                    setupFiles: [".storybook/vitest.setup.ts"],
-                },
-            },
-            {
-                extends: true,
-                // as any is workaround for https://github.com/davidmyersdev/vite-plugin-node-polyfills/issues/150
-                plugins: [nodePolyfills({ include: ["util"], globals: { global: false } }) as any],
-                test: {
-                    name: "unit",
-                    browser: {
-                        enabled: true,
-                        headless: true,
-                        provider: playwright({
-                            // These tests don't actually take screenshots (at least at time of writing)
-                            // but let's pass these options everywhere for consistency
-                            contextOptions: commonContextOptions,
-                            launchOptions: commonLaunchOptions,
-                        }),
-                        instances: [{ browser: "chromium" }],
-                    },
-                    setupFiles: ["src/test/setupTests.ts"],
-                },
-                css: {
-                    modules: {
-                        // Stabilise snapshots while keeping names distinct across CSS modules.
-                        generateScopedName: "[name]_[local]",
+                        setupFiles: [".storybook/vitest.setup.ts"],
                     },
                 },
-            },
-        ],
-    },
-    optimizeDeps: {
-        include: [
-            "vite-plugin-node-polyfills/shims/buffer",
-            "vite-plugin-node-polyfills/shims/process",
-            "@vector-im/compound-design-tokens/assets/web/icons",
-        ],
-    },
-    resolve: {
-        alias: {
-            "@test-utils": path.resolve(__dirname, "./src/test/utils/index.tsx"),
+                {
+                    extends: true,
+                    // as any is workaround for https://github.com/davidmyersdev/vite-plugin-node-polyfills/issues/150
+                    plugins: [nodePolyfills({ include: ["util"], globals: { global: false } }) as any],
+                    test: {
+                        name: "unit",
+                        browser: {
+                            enabled: true,
+                            headless: true,
+                            provider: playwright({
+                                // These tests don't actually take screenshots (at least at time of writing)
+                                // but let's pass these options everywhere for consistency
+                                contextOptions: commonContextOptions,
+                                launchOptions: commonLaunchOptions,
+                            }),
+                            instances: [{ browser: "chromium" }],
+                        },
+                        setupFiles: ["src/test/setupTests.ts"],
+                    },
+                    css: {
+                        modules: {
+                            // Stabilise snapshots while keeping names distinct across CSS modules.
+                            generateScopedName: "[name]_[local]",
+                        },
+                    },
+                },
+            ],
         },
-    },
-});
+        optimizeDeps: {
+            include: [
+                "vite-plugin-node-polyfills/shims/buffer",
+                "vite-plugin-node-polyfills/shims/process",
+                "@vector-im/compound-design-tokens/assets/web/icons",
+            ],
+        },
+        resolve: {
+            alias: {
+                "@test-utils": path.resolve(__dirname, "./src/test/utils/index.tsx"),
+            },
+        },
+    }),
+    true,
+);
